Delete dead screens, models, and the unused KeyboardShortcuts hookup
Removes files with no references outside themselves: the parallel
onboarding PermissionsView, HistoryDetailView, HotkeyConfiguration
(which claimed a Ctrl+Shift+Space default the real hotkey logic never
used), RecordingSession, TranscriptionState, and the never-called
WhisperService.transcribeChunk left over from removed chunk stitching.
AudioInputView shrinks to just DeviceRow, the one component SettingsView
actually uses. The KeyboardShortcuts package was imported in three files
but its API never called and its one shortcut name never consumed —
imports and README mention dropped (package ref left in the project for
a separate pass).
