+++ speaktype/Constants/Shortcuts+Extensions.swift
@@ -1,13 +1,6 @@
 import Foundation
-import KeyboardShortcuts
-import AppKit
-
-extension KeyboardShortcuts.Name {
-    static let toggleRecord = Self("toggleRecord", default: .init(.space, modifiers: [.control, .option]))
-}
 
 extension Notification.Name {
-    static let hotkeyTriggered = Notification.Name("hotkeyTriggered") // Legacy, can be removed
     static let recordingStartRequested = Notification.Name("recordingStartRequested")
     static let recordingStopRequested = Notification.Name("recordingStopRequested")
     static let recordingCancelRequested = Notification.Name("recordingCancelRequested")
